requirements_django.txt
# ===== DJANGO CORE =====
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# ===== AUTENTICACION =====
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0

# ===== BASE DE DATOS =====
psycopg2-binary>=2.9.9  # PostgreSQL (para producción)
# SQLite viene incluido con Python

# ===== TAREAS ASINCRONAS (para ML futuro) =====
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# ===== MACHINE LEARNING (futuro) =====
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.26.2
# TensorFlow y PyTorch se agregarán cuando sea necesario

# ===== VISUALIZACION Y EXPORTACION =====
matplotlib>=3.8.2
plotly>=5.18.0
reportlab>=4.0.7  # Para PDFs
openpyxl>=3.1.2   # Para Excel
Pillow>=10.1.0    # Para imágenes

# ===== DESARROLLO =====
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-django>=4.7.0
black>=23.12.1
flake8>=6.1.0

# ===== SERVIDOR =====
gunicorn>=21.2.0  # Para producción
whitenoise>=6.6.0  # Para servir static files

# ===== ADMIN MEJORADO =====
django-admin-interface>=0.26.0
django-import-export>=3.3.3  # Importar/Exportar datos en admin

# ===== UTILIDADES =====
python-decouple>=3.8  # Alternativa a python-dotenv
django-extensions>=3.2.3  # Comandos útiles
